Update error message localization keys (#1003)
* Update error message localization keys

Replaced old error message and title localization keys with new ones (e.g., 'ErrorMessageBox' to 'ErrorMessageBox_Content', 'errorOcurredTitle' to 'ErrorOccurred_Title') in ErrorService and Updater. This improves consistency and correctness in localized error dialogs.

* Fix typos in error messages and exception texts

Corrected several spelling mistakes in error messages and exception classes, such as 'ocurred' to 'occurred', 'been' to 'be', and 'retrievbed' to 'retrieved', to improve clarity and professionalism.

* fix typo in resource files

and remove unused data